Unify Reducer.presenting API
Mark functions that takes key path or case path as `toLocalState`
parameter. Use function that takes `ReducerPresentingToLocalState`
as `toLocalState` parameter.
