+++ Example/Example/NavigationLinkExample.swift
@@ -29,7 +29,7 @@ struct NavigationLinkExample: View {
   }
     .presenting(
       detailReducer,
-      state: \.detail,
+      state: .keyPath(\.detail),
       action: /MasterAction.detail,
       environment: { () }
     )
