+++ Example/Example/SheetExample.swift
@@ -32,7 +32,7 @@ struct SheetExample: View {
     }
   }.presenting(
     detailReducer,
-    state: \.detail,
+    state: .keyPath(\.detail),
     action: /MasterAction.detail,
     environment: { () }
   )
