+++ Sources/ComposablePresentation/Reducer+Presenting.swift
@@ -1,20 +1,6 @@
 import ComposableArchitecture
 
 extension Reducer {
-  /// Describes presentation action, like `onPresent` or `onDismiss`.
-  public struct ReducerPresentingAction {
-    public typealias Run = (inout State, Environment) -> Effect<Action, Never>
-
-    /// An action that performs no state mutations and returns no effects.
-    public static var empty: Self { .init { _, _ in .none } }
-
-    public init(run: @escaping Run) {
-      self.run = run
-    }
-
-    public var run: Run
-  }
-
   /// Combines the reducer with a local reducer that works on optionally presented `LocalState`.
   ///
   /// - All effects returned by the local reducer will be canceled when `LocalState` becomes `nil`.
@@ -22,7 +8,7 @@ extension Reducer {
   ///
   /// - Parameters:
   ///   - localReducer: A reducer that works on `LocalState`, `LocalAction`, `LocalEnvironment`.
-  ///   - toLocalState: A key path that can get/set `LocalState` inside `State`.
+  ///   - toLocalState: `ReducerPresentingToLocalState` that can get/set `LocalState` inside `State`.
   ///   - toLocalAction: A case path that can extract/embed `LocalAction` from `Action`.
   ///   - toLocalEnvironment: A function that transforms `Environment` into `LocalEnvironment`.
   ///   - onPresent: An action run when `LocalState` is set to an honest value. Defaults to an empty action.
@@ -33,74 +19,11 @@ extension Reducer {
   /// - Returns: A single, combined reducer.
   public func presenting<LocalState, LocalAction, LocalEnvironment>(
     _ localReducer: Reducer<LocalState, LocalAction, LocalEnvironment>,
-    state toLocalState: WritableKeyPath<State, LocalState?>,
-    action toLocalAction: CasePath<Action, LocalAction>,
-    environment toLocalEnvironment: @escaping (Environment) -> LocalEnvironment,
-    onPresent: ReducerPresentingAction = .empty,
-    onDismiss: ReducerPresentingAction = .empty,
-    breakpointOnNil: Bool = true,
-    file: StaticString = #fileID,
-    line: UInt = #line
-  ) -> Self {
-    presenting(
-      localReducer,
-      state: .keyPath(toLocalState),
-      action: toLocalAction,
-      environment: toLocalEnvironment,
-      onPresent: onPresent,
-      onDismiss: onDismiss,
-      breakpointOnNil: breakpointOnNil,
-      file: file,
-      line: line
-    )
-  }
-
-  /// Combines the reducer with a local reducer that works on optionally presented `LocalState`.
-  ///
-  /// - All effects returned by the local reducer will be canceled when `LocalState` becomes `nil`.
-  ///
-  /// - Parameters:
-  ///   - localReducer: A reducer that works on `LocalState`, `LocalAction`, `LocalEnvironment`.
-  ///   - toLocalState: A case path that can extract/embed `LocalState` from `State`.
-  ///   - toLocalAction: A case path that can extract/embed `LocalAction` from `Action`.
-  ///   - toLocalEnvironment: A function that transforms `Environment` into `LocalEnvironment`.
-  ///   - onPresent: An action run when `LocalState` is set to an honest value. Defaults to an empty action.
-  ///   - onDismiss: An action run when `LocalState` becomes `nil`. Defaults to an empty action.
-  ///   - breakpointOnNil: If `true`, raises `SIGTRAP` signal when an action is sent to the reducer
-  ///       but state is in invalid case. This is generally considered a logic error, as a child reducer cannot
-  ///       process a child action for unavailable child state. Default value is `true`.
-  /// - Returns: A single, combined reducer.
-  public func presenting<LocalState, LocalAction, LocalEnvironment>(
-    _ localReducer: Reducer<LocalState, LocalAction, LocalEnvironment>,
-    state toLocalState: CasePath<State, LocalState>,
-    action toLocalAction: CasePath<Action, LocalAction>,
-    environment toLocalEnvironment: @escaping (Environment) -> LocalEnvironment,
-    onPresent: ReducerPresentingAction = .empty,
-    onDismiss: ReducerPresentingAction = .empty,
-    breakpointOnNil: Bool = true,
-    file: StaticString = #fileID,
-    line: UInt = #line
-  ) -> Self {
-    presenting(
-      localReducer,
-      state: .casePath(toLocalState),
-      action: toLocalAction,
-      environment: toLocalEnvironment,
-      onPresent: onPresent,
-      onDismiss: onDismiss,
-      breakpointOnNil: breakpointOnNil,
-      file: file,
-      line: line
-    )
-  }
-
-  func presenting<LocalState, LocalAction, LocalEnvironment>(
-    _ localReducer: Reducer<LocalState, LocalAction, LocalEnvironment>,
     state toLocalState: ReducerPresentingToLocalState<State, LocalState>,
     action toLocalAction: CasePath<Action, LocalAction>,
     environment toLocalEnvironment: @escaping (Environment) -> LocalEnvironment,
-    onPresent: ReducerPresentingAction = .empty,
-    onDismiss: ReducerPresentingAction = .empty,
+    onPresent: ReducerPresentingAction<State, Action, Environment> = .empty,
+    onDismiss: ReducerPresentingAction<State, Action, Environment> = .empty,
     breakpointOnNil: Bool = true,
     file: StaticString = #fileID,
     line: UInt = #line
@@ -113,17 +36,35 @@ extension Reducer {
       let shouldRunLocal = toLocalAction.extract(from: action) != nil
       let localEffects: Effect<Action, Never>
       if shouldRunLocal {
-        localEffects = localReducer
-          .pullback(
-            state: toLocalState,
-            action: toLocalAction,
-            environment: toLocalEnvironment,
-            breakpointOnNil: breakpointOnNil,
-            file: file,
-            line: line
-          )
-          .run(&state, action, environment)
-          .cancellable(id: localEffectsId)
+        switch toLocalState {
+        case let .keyPath(keyPath):
+          localEffects = localReducer
+            .optional(
+              breakpointOnNil: breakpointOnNil,
+              file: file,
+              line: line
+            )
+            .pullback(
+              state: keyPath,
+              action: toLocalAction,
+              environment: toLocalEnvironment
+            )
+            .run(&state, action, environment)
+            .cancellable(id: localEffectsId)
+
+        case let .casePath(casePath):
+          localEffects = localReducer
+            .pullback(
+              state: casePath,
+              action: toLocalAction,
+              environment: toLocalEnvironment,
+              breakpointOnNil: breakpointOnNil,
+              file: file,
+              line: line
+            )
+            .run(&state, action, environment)
+            .cancellable(id: localEffectsId)
+        }
       } else {
         localEffects = .none
       }
@@ -151,45 +92,20 @@ extension Reducer {
       )
     }
   }
-
-  func pullback<GlobalState, GlobalAction, GlobalEnvironment>(
-    state toLocalState: ReducerPresentingToLocalState<GlobalState, State>,
-    action toLocalAction: CasePath<GlobalAction, Action>,
-    environment toLocalEnvironment: @escaping (GlobalEnvironment) -> Environment,
-    breakpointOnNil: Bool,
-    file: StaticString = #fileID,
-    line: UInt = #line
-  ) -> Reducer<GlobalState, GlobalAction, GlobalEnvironment> {
-    switch toLocalState {
-    case let .keyPath(keyPath):
-      return optional(
-        breakpointOnNil: breakpointOnNil,
-        file: file,
-        line: line
-      ).pullback(
-        state: keyPath,
-        action: toLocalAction,
-        environment: toLocalEnvironment
-      )
-
-    case let .casePath(casePath):
-      return pullback(
-        state: casePath,
-        action: toLocalAction,
-        environment: toLocalEnvironment,
-        breakpointOnNil: breakpointOnNil,
-        file: file,
-        line: line
-      )
-    }
-  }
 }
 
-enum ReducerPresentingToLocalState<State, LocalState> {
+/// `State` ↔ `LocalState` transformation for `.presenting` higher order reducer.
+public enum ReducerPresentingToLocalState<State, LocalState> {
+  /// A key path that can get/set `LocalState` inside `State`.
   case keyPath(WritableKeyPath<State, LocalState?>)
+
+  /// A case path that can extract/embed `LocalState` from `State`.
   case casePath(CasePath<State, LocalState>)
 
-  func callAsFunction(_ state: State) -> LocalState? {
+  /// Returns optional `LocalState` from provided `State`.
+  /// - Parameter state: `State`
+  /// - Returns: Optional `LocalState`
+  public func callAsFunction(_ state: State) -> LocalState? {
     switch self {
     case let .keyPath(keyPath):
       return state[keyPath: keyPath]
@@ -197,6 +113,20 @@ enum ReducerPresentingToLocalState<State, LocalState> {
       return casePath.extract(from: state)
     }
   }
+}
+
+/// Describes presentation action, like `onPresent` or `onDismiss`.
+public struct ReducerPresentingAction<State, Action, Environment> {
+  public typealias Run = (inout State, Environment) -> Effect<Action, Never>
+
+  /// An action that performs no state mutations and returns no effects.
+  public static var empty: Self { .init { _, _ in .none } }
+
+  public init(run: @escaping Run) {
+    self.run = run
+  }
+
+  public var run: Run
 }
 
 struct ReducerPresentingEffectId: Hashable {
