+++ Tests/ComposablePresentationTests/ReducerPresentingTests.swift
@@ -4,8 +4,128 @@ import ComposableArchitecture
 import XCTest
 @testable import ComposablePresentation
 
-final class ReducerPresentingCasePathTests: XCTestCase {
-  func testPresenting() {
+final class ReducerPresentingTests: XCTestCase {
+  func testPresentingWithKeyPath() {
+    var didPresentDetail = 0
+    var didRunDetailReducer = 0
+    var didFireDetailEffect = 0
+    var didDismissDetail = 0
+    var didCancelDetailEffect = 0
+
+    struct MasterState: Equatable {
+      var detail: DetailState?
+    }
+
+    enum MasterAction: Equatable {
+      case presentDetail
+      case dismissDetail
+      case detail(DetailAction)
+    }
+
+    struct MasterEnvironment {
+      var detail: DetailEnvironment
+    }
+
+    let masterReducer = Reducer<MasterState, MasterAction, MasterEnvironment>
+    { state, action, env in
+      switch action {
+      case .presentDetail:
+        state.detail = DetailState()
+        return .none
+
+      case .dismissDetail:
+        state.detail = nil
+        return .none
+
+      case .detail:
+        return .none
+      }
+    }
+
+    struct DetailState: Equatable {}
+
+    enum DetailAction: Equatable {
+      case performEffect
+      case didPerformEffect
+    }
+
+    struct DetailEnvironment {
+      var effect: () -> Effect<Void, Never>
+    }
+
+    let detailReducer = Reducer<DetailState, DetailAction, DetailEnvironment>
+    { state, action, env in
+      didRunDetailReducer += 1
+      switch action {
+      case .performEffect:
+        return env.effect()
+          .map { _ in DetailAction.didPerformEffect }
+          .eraseToEffect()
+
+      case .didPerformEffect:
+        return .none
+      }
+    }
+
+    let store = TestStore(
+      initialState: MasterState(),
+      reducer: masterReducer
+        .presenting(
+          detailReducer,
+          state: .keyPath(\.detail),
+          action: /MasterAction.detail,
+          environment: \.detail,
+          onPresent: .init { _, _ in
+            didPresentDetail += 1
+            return .none
+          },
+          onDismiss: .init { _, _ in
+            didDismissDetail += 1
+            return .none
+          }
+        ),
+      environment: MasterEnvironment(
+        detail: DetailEnvironment(effect: {
+          Empty(completeImmediately: false)
+            .handleEvents(
+              receiveSubscription: { _ in didFireDetailEffect += 1 },
+              receiveCancel: { didCancelDetailEffect += 1 }
+            )
+            .eraseToEffect()
+        })
+      )
+    )
+
+    store.send(.presentDetail) {
+      $0.detail = DetailState()
+    }
+
+    XCTAssertEqual(didPresentDetail, 1)
+    XCTAssertEqual(didRunDetailReducer, 0)
+    XCTAssertEqual(didFireDetailEffect, 0)
+    XCTAssertEqual(didDismissDetail, 0)
+    XCTAssertEqual(didCancelDetailEffect, 0)
+
+    store.send(.detail(.performEffect))
+
+    XCTAssertEqual(didPresentDetail, 1)
+    XCTAssertEqual(didRunDetailReducer, 1)
+    XCTAssertEqual(didFireDetailEffect, 1)
+    XCTAssertEqual(didDismissDetail, 0)
+    XCTAssertEqual(didCancelDetailEffect, 0)
+
+    store.send(.dismissDetail) {
+      $0.detail = nil
+    }
+
+    XCTAssertEqual(didPresentDetail, 1)
+    XCTAssertEqual(didRunDetailReducer, 1)
+    XCTAssertEqual(didFireDetailEffect, 1)
+    XCTAssertEqual(didDismissDetail, 1)
+    XCTAssertEqual(didCancelDetailEffect, 1)
+  }
+
+  func testPresentingWithCasePath() {
     var didPresentFirst = 0
     var didRunFirstReducer = 0
     var didFireFirstEffect = 0
@@ -106,7 +226,7 @@ final class ReducerPresentingCasePathTests: XCTestCase {
       reducer: masterReducer
         .presenting(
           firstReducer,
-          state: /MasterState.first,
+          state: .casePath(/MasterState.first),
           action: /MasterAction.first,
           environment: \.first,
           onPresent: .init { _, _ in
@@ -120,7 +240,7 @@ final class ReducerPresentingCasePathTests: XCTestCase {
         )
         .presenting(
           secondReducer,
-          state: /MasterState.second,
+          state: .casePath(/MasterState.second),
           action: /MasterAction.second,
           environment: \.second,
           onPresent: .init { _, _ in
